First (public) commit!
Don't worry, the history prior to this commit really isn't very interesting or helpful ;)

Scalars and Vectors mostly work.
Assigning doesn't really work yet.
Matrices need a lot of work.

In general, a lot of things still need a lot of work, but this code serves as a great starting point!
It's well structured and has pretty reasonable tests.